backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4

# LLM Providers
google-generativeai==0.3.0
cohere==5.0.0
openai==1.3.0

# LangChain (for RAG only)
langchain==0.1.0
langchain-openai==0.0.5
langchain-postgres==0.0.6
langchain-community==0.0.13

# Template Engine
jinja2==3.1.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Logging
colorlog==6.7.0

# Utilities
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0